Add dashboard for per-user pod diagnostics
"My kernel is crashing" is one of the most common
complaints when supporting users on a JupyterHub, and
figuring out 'how much memory is the user who has problems using?'
is the primary diagnostic step. This commit adds a dashboard
that lets us answer that question for memory & CPU very easily.

This is specific to user *pods*, so each user server in a hub
with many named servers will get its own entry.