Replace global dashboard with activity dashboard
The activity.jsonnet dashboard is extracting four panels from the
jupyterhub.jsonnet dashboard. The running servers panel and
daily/weekly/monhtly active users panels.

By doing this, we get a dashboard with only those four panels, grouped
into a row. This row is then made repeatable over the prometheus
datasource variable. With a repeating row, we could end up with very
many rows, making any panels below hard to get to. Due to that, it makes
sense to use repeating rows in a dedicated dashboard.

The global dashboard that is being deleted made use of hardcoded state
in the grafana instance it was being deployed to. This then also
required regular re-deploys of the dashboard just to update the
hardcoded entries.

Further, the global dashboard couldn't be pre-rendered and re-used
across grafana instances - because it had grafance instance specific
state within it.

With the new Activity dashboard, we can avoid all this complexity and
possibly publish pre-rendered dashboards that any grafana instance can
install directly.