Remove always failing linter
This was introduced in https://github.com/jupyterhub/grafana-dashboards/pull/106,
and has never passed. I don't think we should have a test that is always
expected to fail.

If it should be re-introduced, it should be done in such a way that
whatever changes are needed to make it pass should be part of the PR.

This only reverts the linter, but keeps the rendering test - as that would
help us catch possible jsonnet syntax issues.